+++ HISTORY.rst
@@ -1,5 +1,20 @@
 Release History
 ^^^^^^^^^^^^^^^
+2.0.0 (2015-03-04)
+++++++++++++++++++
+
+* 2.0 Version bump due to incompatible API changes relating to the location of the trakt api_key attribute
+* Add additional debug logging for API responses
+* Add tmdb_id to the `TVShow.ids` attribute
+* Fixed `trakt.init` to instruct users on how to create a new OAuth application
+* * Fixed `TVSeason.to_json` to return accurately scoped season information
+* Updated documentation on APIv2's Authentication patterns
+
+1.0.3 (2015-02-28)
+++++++++++++++++++
+
+* Fixed a bug with `First Aired Date` datetime parsing
+
 1.0.2 (2015-02-17)
 ++++++++++++++++++